src/code/serial-at-api/requirements.txt
Flask
pyserial
